Include inlined messages in the method literal frames
Some (typically control) messages are inlined for perf reasons,
e.g. #whileTrue:. Historically Dolphin's compiler did not store the
inlined message selector in the literal frame of sending methods to save
space, but this makes searching for references to these a special case
that requires parsing potential senders. This space optimisation seems
inappropriate today, and isn't worth the extra complexity in the IDE.