Weak objects losing refs should not be sent loss notifications if garbage
Weak objects that are themselves garbage should not be rescued and sent
notifications of loss as this may result in a very large number of
GC cycles being required to collect large object structures using (for
example) event registration. Since the purpose of the notifications is to
allow objects to keep their internal data structures consistent, there is
no point carrying out repairs when the weak object itself is garbage.
The exception is finalizable weak objects - these will be rescued and
placed in the finalization queue, and will also receive a weak-loss
notification in the GC cycle in which they are finalized.