Compiler eval of simple const expressions can avoid dispatch to Smalltalk
STL uses ## expressions a lot for references to classes embedded in the
array of literals, and this makes compiling the resource methods slow.
This is a simple optimisation to speed up this particular case by avoiding
the unecessary callback into Smalltalk to execute expressions where the
result is already available to the compiler.