Bug in oopq prevents image restart if undertaker queue not empty
The VM's OopQueue class wraps the circular buffers used for the
memory manager finalization and weak notification queues. It has to
re-establish the head and tail of the queue on image load, but has an
off-by-one error for non-empty queues and includes an empty slot (which
will contain nil) as one of the queue elements. This is benign for the
finalization queue, but causes a "nil DNU #elementsExpired:" error for the
weak notifications queue.