Add FP primitives to take advantage of SSE2 intrinsics
Many of the FP operations call the CRT via external calls from the image.
This decision was trading off performance against having to maintain a
large number of FP primitives written in assembler. Now that the FP
primitives can be written efficiently in C++, there is no reason not to
add primitives for the common mathematical functions.