Remove the 'at' caches
The at/atput caches are small caches of data about a few recently accessed
objects that allow shortcutting of normal primitive invocation for
element-by-element access through the #at: and #at:put: messages. When
implemented the overhead of maintaining these was fairly small, but in the
current VM they are flushed very frequently. It turns out that the main
beneficial effect of these is when accessing Arrays, and optimising for
these can be achieved much more simply. There might be some benefit in
also optimising for ByteArray.