Reimplement the hashBytes() function in C++
This reproduces the existing algorithm, including the poor results on some
sequences such as #[255 255 255 255 255 255 16], which yields a hash of 16,
losing all bits from the other elements. This is due to the way the carry
is handled. I think it is probably OK to fix, as it doesn't affect string
hashing of 7-bit ascii characters.