Replace some x86 asm SmallInteger prims with C++ implementations
Includes C++ implementations for the following primitives:
primitiveBitAnd
primitiveBitOr
primitiveBitXor
primitiveAllMask
primitiveAnyMask

The SmallInteger arithmetic primitives might not be implementable in C++
with the same efficiency as in asm because of limitations of C in
detecting overflow and similar conditions. Unless Intel intrinsics can
help, these may need to remain in assembler.

There are a few more integer primitives that can be switched to C++ without
a perf hit, however. These will follow.

Also moved a load of system includes to the precompiled header. This should
help speed up the compilation.

Contributes to #148